Added Compensation Log view. Fixed PlayerID not being logged in CompensationLogs. Laid groundwork for basic role management.

+++ RGNCompensation.Models/CompensationLog.cs
@@ -9,7 +9,7 @@ namespace RGNCompensation.Models
     {
         public int reimburse_id { get; set; }
         public int reimburse_amount { get; set; }
-        public int reimburse_player_id { get; set; }
+        public double reimburse_player_id { get; set; }
         public string reimburse_justification { get; set; }
         public string reimburse_admin { get; set; }
         //public string reimburse_timestamp { get; set; }
